Configuração da API de envio de emails + Verificação de conta por código

+++ src/address/address.service.ts
@@ -1,18 +1,29 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { CreateAddressDto } from './dto/create-address.dto';
 import { UpdateAddressDto } from './dto/update-address.dto';
 import { Address } from './interfaces/address.interface';
+import { Ride } from '@src/ride/interfaces/ride.interface';
 
 @Injectable()
 export class AddressService {
-  constructor(@InjectModel('Address') private addressModel: Model<Address>) {}
+  constructor(
+    @InjectModel('Address') private addressModel: Model<Address>,
+    @InjectModel('Ride') private rideModel: Model<Ride>,
+  ) {}
 
-  async create(createAddressDto: CreateAddressDto, userId: string): Promise<Address> {
-    const newAddress = {...createAddressDto, user: userId}
+  async create(
+    createAddressDto: CreateAddressDto,
+    userId: string,
+  ): Promise<Address> {
+    const newAddress = { ...createAddressDto, user: userId };
     const address = new this.addressModel(newAddress);
-    
+
     return await address.save();
   }
 
@@ -28,8 +39,13 @@ export class AddressService {
     return address;
   }
 
-  async update(id: string, updateAddressDto: UpdateAddressDto): Promise<Address> {
-    const updatedAddress = await this.addressModel.findByIdAndUpdate(id, updateAddressDto, { new: true }).exec();
+  async update(
+    id: string,
+    updateAddressDto: UpdateAddressDto,
+  ): Promise<Address> {
+    const updatedAddress = await this.addressModel
+      .findByIdAndUpdate(id, updateAddressDto, { new: true })
+      .exec();
     if (!updatedAddress) {
       throw new NotFoundException(`Endereço com ID ${id} não encontrado`);
     }
@@ -37,11 +53,16 @@ export class AddressService {
   }
 
   async remove(id: string): Promise<Address> {
+    const exists = await this.rideModel.findOne({ car: id, isOver: false });
+    if (exists)
+      throw new BadRequestException(
+        'O endereço está sendo utilizado em uma carona e não pode ser removido.',
+      );
     const result = await this.addressModel.findByIdAndDelete(id).exec();
     if (!result) {
       throw new NotFoundException(`Endereço com ID ${id} não encontrado`);
     }
-    return result
+    return result;
   }
 
   async removeAll(): Promise<void> {
@@ -49,7 +70,7 @@ export class AddressService {
   }
 
   async getUserAddresses(userId: string): Promise<Address[]> {
-    const userAddress = await this.addressModel.find({user: userId}).exec()
-    return userAddress
+    const userAddress = await this.addressModel.find({ user: userId }).exec();
+    return userAddress;
   }
 }
